refactor: Change BM25 vector type from f32 to u32

+++ src/bm25_embedder.rs
@@ -1,7 +1,7 @@
-use bm25::{DefaultTokenizer, Tokenizer, EmbedderBuilder};
+use bm25::{DefaultTokenizer, EmbedderBuilder};
 
-pub fn create_bm25_vector(text: &str) -> Vec<f32> {
-    let embedder = EmbedderBuilder::<f32, DefaultTokenizer>::with_avgdl(1.0).build();
+pub fn create_bm25_vector(text: &str) -> Vec<u32> {
+    let embedder = EmbedderBuilder::<u32, DefaultTokenizer>::with_avgdl(1.0).build();
     let embedding = embedder.embed(text);
     embedding.indices().cloned().collect::<Vec<_>>()
 }
